requirements.txt
bitstring==4.0.1
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.0.4
colorama==0.4.6
construct==2.10.68
contextlib2==21.6.0
contourpy==1.0.7
cryptography==36.0.2
cycler==0.11.0
ecdsa==0.18.0
esp-coredump==1.5.0
esptool==4.5.1
fonttools==4.39.3
freertos-gdb==1.0.2
future==0.18.2
idf-component-manager==1.2.2
idna==3.4
kconfiglib==14.1.0
kiwisolver==1.4.4
lockfile==0.12.2
matplotlib==3.7.1
msgpack==1.0.5
numpy==1.24.2
packaging==23.0
Pillow==9.5.0
pycparser==2.21
pyelftools==0.27
pygdbmi==0.9.0.2
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.0
PyQtChart==5.15.6
PyQtChart-Qt5==5.15.2
pyserial==3.5
python-dateutil==2.8.2
PyYAML==6.0
reedsolo==1.6.0
requests==2.28.2
requests-file==1.5.1
requests-toolbelt==0.10.1
schema==0.7.5
six==1.16.0
tqdm==4.65.0
urllib3==1.26.15
